Adiconando funcao remover imcompleta
